Add files via upload
This has a started page for the ticket forms.  It is not complete.  I am not aware of the way to connect the page to the rest of the project properly, but the page and button are there for whoever wants to tackle connection.  I plan to populate the form more fully and make it conform to the common theme the next chance I get.