Add files via upload
Not sure why the commit didn't work, but this is the version that I cleaned up code on and inserted the Ticket Screen, even though the Ticket Screen itself is not working with the rest of the project yet.  Feel free to start from this version, as the rest of the project is stable. and more easily read.